+++ sdk/resourcemanager/healthdataaiservices/armhealthdataaiservices/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.1.0 (2026-01-30)
+### Features Added
+
+- New enum type `DeidServiceSKUName` with values `DeidServiceSKUNameBasic`, `DeidServiceSKUNameFree`, `DeidServiceSKUNameStandard`
+- New enum type `SKUTier` with values `SKUTierBasic`, `SKUTierFree`, `SKUTierPremium`, `SKUTierStandard`
+- New struct `DeidServiceSKU`
+- New field `SKU` in struct `DeidService`
+- New field `SKU` in struct `DeidUpdate`
+
+
 ## 1.0.0 (2024-11-20)
 ### Breaking Changes